The great refactor part6
