+++ api/src/main/java/cz/metacentrum/perun/spRegistration/rest/controllers/AppErrorController.java
@@ -1,4 +1,4 @@
-package cz.metacentrum.perun.spRegistration.rest.controllers.common;
+package cz.metacentrum.perun.spRegistration.rest.controllers;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.web.servlet.error.ErrorController;
