+++ api/src/main/java/cz/metacentrum/perun/spRegistration/rest/controllers/AuthController.java
@@ -1,4 +1,4 @@
-package cz.metacentrum.perun.spRegistration.rest.controllers.common;
+package cz.metacentrum.perun.spRegistration.rest.controllers;
 
 import cz.metacentrum.perun.spRegistration.common.models.User;
 import lombok.extern.slf4j.Slf4j;
