+++ api/src/main/java/cz/metacentrum/perun/spRegistration/rest/controllers/FacilitiesController.java
@@ -1,11 +1,14 @@
-package cz.metacentrum.perun.spRegistration.rest.controllers.facilities;
+package cz.metacentrum.perun.spRegistration.rest.controllers;
 
+
+import cz.metacentrum.perun.spRegistration.common.enums.AttributeCategory;
 import cz.metacentrum.perun.spRegistration.common.exceptions.CodeNotStoredException;
 import cz.metacentrum.perun.spRegistration.common.exceptions.ExpiredCodeException;
 import cz.metacentrum.perun.spRegistration.common.exceptions.InternalErrorException;
 import cz.metacentrum.perun.spRegistration.common.exceptions.UnauthorizedActionException;
 import cz.metacentrum.perun.spRegistration.common.models.Facility;
 import cz.metacentrum.perun.spRegistration.common.models.LinkCode;
+import cz.metacentrum.perun.spRegistration.common.models.PerunAttribute;
 import cz.metacentrum.perun.spRegistration.common.models.User;
 import cz.metacentrum.perun.spRegistration.persistence.models.ProvidedService;
 import cz.metacentrum.perun.spRegistration.persistence.exceptions.PerunConnectionException;
@@ -13,8 +16,10 @@ import cz.metacentrum.perun.spRegistration.persistence.exceptions.PerunUnknownEx
 import cz.metacentrum.perun.spRegistration.rest.ApiUtils;
 import cz.metacentrum.perun.spRegistration.service.AddAdminsService;
 import cz.metacentrum.perun.spRegistration.service.FacilitiesService;
+import cz.metacentrum.perun.spRegistration.service.UtilsService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -29,21 +34,54 @@ import java.security.InvalidKeyException;
 import java.util.List;
 
 /**
- * Controller handling USER actions related to Facilities.
+ * Controller handling actions related to Facilities.
  *
  * @author Dominik Frantisek Bucik <bucik@ics.muni.cz>;
  */
 @RestController
 @Slf4j
-public class UserFacilitiesController {
+public class FacilitiesController {
 
 	private final FacilitiesService facilitiesService;
 	private final AddAdminsService addAdminsService;
+	private final UtilsService utilsService;
 
 	@Autowired
-	public UserFacilitiesController(FacilitiesService facilitiesService, AddAdminsService addAdminsService) {
+	public FacilitiesController(FacilitiesService facilitiesService,
+								AddAdminsService addAdminsService,
+								UtilsService utilsService)
+	{
 		this.addAdminsService = addAdminsService;
 		this.facilitiesService = facilitiesService;
+		this.utilsService = utilsService;
+	}
+
+	@GetMapping(path = "/api/facility/{facilityId}")
+	public Facility facilityDetail(@SessionAttribute("user") User user,
+								   @PathVariable("facilityId") Long facilityId)
+			throws UnauthorizedActionException, InternalErrorException, BadPaddingException, InvalidKeyException, IllegalBlockSizeException, PerunUnknownException, PerunConnectionException {
+		log.trace("facilityDetail(user(): {}, facilityId: {})", user.getId(), facilityId);
+
+		Facility facility = facilitiesService.getFacility(facilityId, user.getId(), true, true);
+
+		log.trace("facilityDetail() returns: {}", facility);
+		return facility;
+	}
+
+	@GetMapping(path = "/api/facility/signature/{facilityId}")
+	public Facility facilityDetailSignature(@SessionAttribute("user") User user,
+											@PathVariable("facilityId") Long facilityId)
+			throws UnauthorizedActionException, InternalErrorException, BadPaddingException, InvalidKeyException,
+			IllegalBlockSizeException, PerunUnknownException, PerunConnectionException
+	{
+		log.trace("facilityDetailSignature(user(): {}, facilityId: {})", user.getId(), facilityId);
+
+		Facility facility = facilitiesService.getFacility(facilityId, user.getId(), false, false);
+		facility.getAttributes().get(AttributeCategory.PROTOCOL).clear();
+		facility.getAttributes().get(AttributeCategory.ACCESS_CONTROL).clear();
+
+		log.trace("facilityDetailSignature() returns: {}", facility);
+		return facility;
 	}
 
 	@GetMapping(path = "/api/userFacilities")
@@ -92,7 +130,7 @@ public class UserFacilitiesController {
 
 	@PostMapping(path = "/api/addAdmin/reject")
 	public void addAdminReject(@SessionAttribute("user") User user,
-								   @RequestBody String code)
+							   @RequestBody String code)
 			throws BadPaddingException, ExpiredCodeException, IllegalBlockSizeException,
 			InvalidKeyException, InternalErrorException, CodeNotStoredException {
 		log.trace("addAdminReject(user: {}, code: {})", user, code);
@@ -139,6 +177,28 @@ public class UserFacilitiesController {
 
 		log.trace("getDetailedFacilityWithInputs() returns: {}", facility);
 		return facility;
+	}
+
+	@GetMapping(path = "/api/allFacilities")
+	public List<ProvidedService> allFacilities(@SessionAttribute("user") User user) throws UnauthorizedActionException, PerunUnknownException, PerunConnectionException {
+		log.trace("allFacilities({})", user.getId());
+
+		List<ProvidedService> facilityList = facilitiesService.getAllFacilities(user.getId());
+
+		log.trace("allFacilities() returns: {}", facilityList);
+		return facilityList;
+	}
+
+	@PostMapping(path = "api/facility/regenerateClientSecret/{facilityId}",
+			produces = MediaType.APPLICATION_JSON_VALUE)
+	public PerunAttribute generateClientSecret(@SessionAttribute("user") User user,
+											   @PathVariable("facilityId") Long facilityId)
+			throws UnauthorizedActionException, BadPaddingException, InvalidKeyException, IllegalBlockSizeException, PerunUnknownException, PerunConnectionException {
+		log.trace("generateClientSecret(user: {}, facilityId: {})", user, facilityId);
+		PerunAttribute clientSecret = utilsService.regenerateClientSecret(user.getId(), facilityId);
+
+		log.trace("generateClientSecret() returns: {}", clientSecret);
+		return clientSecret;
 	}
 
 }
