+++ api/src/main/java/cz/metacentrum/perun/spRegistration/rest/controllers/GlobalExceptionController.java
@@ -1,4 +1,4 @@
-package cz.metacentrum.perun.spRegistration.rest.controllers.common;
+package cz.metacentrum.perun.spRegistration.rest.controllers;
 
 import cz.metacentrum.perun.spRegistration.common.exceptions.CannotChangeStatusException;
 import cz.metacentrum.perun.spRegistration.common.exceptions.CodeNotStoredException;
