+++ api/src/main/java/cz/metacentrum/perun/spRegistration/rest/controllers/HomeController.java
@@ -1,4 +1,4 @@
-package cz.metacentrum.perun.spRegistration.rest.controllers.common;
+package cz.metacentrum.perun.spRegistration.rest.controllers;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
