+++ api/src/main/java/cz/metacentrum/perun/spRegistration/rest/controllers/RequestsController.java
@@ -1,5 +1,6 @@
-package cz.metacentrum.perun.spRegistration.rest.controllers.requests;
+package cz.metacentrum.perun.spRegistration.rest.controllers;
 
+import cz.metacentrum.perun.spRegistration.common.exceptions.ActiveRequestExistsException;
 import cz.metacentrum.perun.spRegistration.common.exceptions.CannotChangeStatusException;
 import cz.metacentrum.perun.spRegistration.common.exceptions.InternalErrorException;
 import cz.metacentrum.perun.spRegistration.common.exceptions.UnauthorizedActionException;
@@ -24,20 +25,96 @@ import java.security.InvalidKeyException;
 import java.util.List;
 
 /**
- * Controller handling ADMIN actions related to Requests.
+ * Controller handling actions related to Requests.
  *
  * @author Dominik Frantisek Bucik <bucik@ics.muni.cz>;
  */
 @RestController
 @Slf4j
-public class AdminRequestsController {
-	
+public class RequestsController {
+
 	private final RequestsService requestsService;
 
 	@Autowired
-	public AdminRequestsController(RequestsService requestsService) {
+	public RequestsController(RequestsService requestsService) {
 		this.requestsService = requestsService;
 	}
+
+	@GetMapping(path = "/api/userRequests")
+	public List<Request> userRequests(@SessionAttribute("user") User user) throws PerunUnknownException, PerunConnectionException {
+		log.trace("userRequests({})", user.getId());
+
+		List<Request> requestList = requestsService.getAllUserRequests(user.getId());
+
+		log.trace("userRequests() returns: {}", requestList);
+		return requestList;
+	}
+
+	@PostMapping(path = "/api/register")
+	public Long createRegistrationRequest(@SessionAttribute("user") User user,
+										  @RequestBody List<PerunAttribute> attributes) throws InternalErrorException
+	{
+		log.trace("createRegistrationRequest(user: {}, attributes: {})", user.getId(), attributes);
+
+		Long generatedId = requestsService.createRegistrationRequest(user.getId(), attributes);
+
+		log.trace("createRegistrationRequest() returns: {}", generatedId);
+		return generatedId;
+	}
+
+	@PostMapping(path = "/api/changeFacility/{facilityId}")
+	public Long createFacilityChangesRequest(@SessionAttribute("user") User user,
+											 @RequestBody List<PerunAttribute> attributes,
+											 @PathVariable("facilityId") Long facilityId)
+			throws ActiveRequestExistsException, InternalErrorException, UnauthorizedActionException, PerunUnknownException, PerunConnectionException {
+		log.trace("createFacilityChangesRequest(user: {}, facilityId: {}, attributes: {})", user.getId(),
+				facilityId, attributes);
+
+		Long generatedId = requestsService.createFacilityChangesRequest(facilityId, user.getId(), attributes);
+
+		log.trace("createFacilityChangesRequest() returns: {}", generatedId);
+		return generatedId;
+	}
+
+	@PostMapping(path = "/api/remove/{facilityId}")
+	public Long createRemovalRequest(@SessionAttribute("user") User user,
+									 @PathVariable("facilityId") Long facilityId)
+			throws ActiveRequestExistsException, InternalErrorException, UnauthorizedActionException, PerunUnknownException, PerunConnectionException {
+		log.trace("createRemovalRequest(user: {}, facilityId: {})", user.getId(), facilityId);
+
+		Long generatedId = requestsService.createRemovalRequest(user.getId(), facilityId);
+
+		log.trace("createRemovalRequest() returns: {}", generatedId);
+		return generatedId;
+	}
+
+	@PostMapping(path = "/api/update/{requestId}")
+	public boolean updateRequest(@SessionAttribute("user") User user,
+								 @PathVariable("requestId") Long requestId,
+								 @RequestBody List<PerunAttribute> attributes)
+			throws InternalErrorException, UnauthorizedActionException
+	{
+		log.trace("updateRequest(user: {}, requestId: {}, attributes: {})", user.getId(), requestId, attributes);
+
+		boolean successful = requestsService.updateRequest(requestId, user.getId(), attributes);
+
+		log.trace("updateRequest() returns: {}", successful);
+		return successful;
+	}
+
+	@GetMapping(path = "/api/request/{requestId}")
+	public Request requestDetail(@SessionAttribute("user") User user,
+								 @PathVariable("requestId") Long requestId)
+			throws InternalErrorException, UnauthorizedActionException, PerunUnknownException, PerunConnectionException {
+		log.trace("requestDetail(user: {}, requestId: {})", user.getId(), requestId);
+
+		Request request = requestsService.getRequest(requestId, user.getId());
+
+		log.trace("requestDetail() returns: {}", request);
+		return request;
+	}
+
+	// admin
 
 	@GetMapping(path = "/api/allRequests")
 	public List<Request> allRequests(@SessionAttribute("user") User user)
@@ -56,7 +133,7 @@ public class AdminRequestsController {
 								  @PathVariable("requestId") Long requestId)
 			throws CannotChangeStatusException, InternalErrorException, UnauthorizedActionException, BadPaddingException, InvalidKeyException, IllegalBlockSizeException, PerunUnknownException, PerunConnectionException {
 		log.trace("approveRequest(user: {}, requestId: {})", user.getId(), requestId);
-		
+
 		boolean successful = requestsService.approveRequest(requestId, user.getId());
 
 		log.trace("approveRequest() returns: {}", successful);
@@ -69,7 +146,7 @@ public class AdminRequestsController {
 			throws UnauthorizedActionException, CannotChangeStatusException, InternalErrorException
 	{
 		log.trace("rejectRequest(user: {}, requestId: {})", user.getId(), requestId);
-		
+
 		boolean successful = requestsService.rejectRequest(requestId, user.getId());
 
 		log.trace("rejectRequest() returns: {}", successful);
@@ -83,7 +160,7 @@ public class AdminRequestsController {
 			throws UnauthorizedActionException, CannotChangeStatusException, InternalErrorException
 	{
 		log.trace("askForChanges(user: {}, requestId: {}, attributes: {})", user.getId(), requestId, attributes);
-		
+
 		boolean successful = requestsService.askForChanges(requestId, user.getId(), attributes);
 
 		log.trace("askForChanges() returns: {}", successful);
